+++ src/main.rs
@@ -10,7 +10,7 @@ mod camera;
 mod util;
 
 use rand::{thread_rng, Rng};
-use std::f64::INFINITY;
+use std::f32::INFINITY;
 use std::fs::File;
 use std::io::prelude::*;
 
@@ -19,7 +19,7 @@ enum Object {
 }
 
 impl Hittable for Object {
-  fn hit(&self, ray: &Ray, tmin: f64, tmax: f64) -> Option<Hit> {
+  fn hit(&self, ray: &Ray, tmin: f32, tmax: f32) -> Option<Hit> {
     match self {
       Object::Sphere(sp) => sp.hit(ray, tmin, tmax),
     }
@@ -27,7 +27,7 @@ impl Hittable for Object {
 }
 
 impl Hittable for Vec<Object> {
-  fn hit(&self, ray: &Ray, tmin: f64, tmax: f64) -> Option<Hit> {
+  fn hit(&self, ray: &Ray, tmin: f32, tmax: f32) -> Option<Hit> {
     self
       .iter()
       .fold(
@@ -50,6 +50,7 @@ fn ray_colour(ray: &Ray, world: &Vec<Object>, depth: u32) -> Vec3 {
     let target = hit.p + hit.norm + Vec3::random_unit_sphere();
     return 0.5 * ray_colour(&ray::ray(hit.p, target - hit.p), world, depth - 1);
   }
+
   let unit_direction = ray.dir.unit();
   let t = 0.5 * (unit_direction.y + 1.0);
 
@@ -60,7 +61,7 @@ fn main() -> std::io::Result<()> {
   // image
   let aspect_ratio = 16.0 / 9.0;
   let image_width = 400;
-  let image_height = (image_width as f64 / aspect_ratio) as u32;
+  let image_height = (image_width as f32 / aspect_ratio) as u32;
 
   let world = vec![
     Object::Sphere(Sphere {
@@ -89,8 +90,8 @@ fn main() -> std::io::Result<()> {
 
     for i in 0..image_width {
       let colour = (0..samples_per_pixel).fold(Vec3::zero(), |colour, _sample| {
-        let u = (i as f64 + rnd.gen_range(0.0..1.0)) / (image_width - 1) as f64;
-        let v = (j as f64 + rnd.gen_range(0.0..1.0)) / (image_height - 1) as f64;
+        let u = (i as f32 + rnd.gen_range(0.0..1.0)) / (image_width - 1) as f32;
+        let v = (j as f32 + rnd.gen_range(0.0..1.0)) / (image_height - 1) as f32;
 
         let ray = camera::SimpleCamera::ray(u, v);
